move default kubernetes scanning as scan subcommand  - fixes #133
