+++ cmd/reliably/scan/shared.go
@@ -1,0 +1,7 @@
+package scan
+
+import (
+	"github.com/reliablyhq/cli/cmd/reliably/cmdutil"
+)
+
+type Choice = cmdutil.Choice
